career_backend/career_navigator/requirements.txt
google.generativeai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
django
djangorestframework
django-cors-headers